+++ sources/packages/web/src/composables/useFormioUtils.ts
@@ -1,3 +1,5 @@
+import { StudentService } from "@/services/StudentService";
+import { StudentUploadFileDto } from "@/types";
 import { Utils } from "formiojs";
 
 /**
@@ -143,6 +145,23 @@ export function useFormioUtils() {
     checkbox.redraw();
   };
 
+  /**
+   * Used to download the document or file uploaded, which has the StudentUploadFileDto structure
+   * @param studentDocument
+   */
+  const downloadDocument = async (studentDocument: StudentUploadFileDto) => {
+    const fileURL = await StudentService.shared.downloadStudentFile(
+      studentDocument.uniqueFileName,
+    );
+    const fileLink = document.createElement("a");
+    fileLink.href = fileURL;
+    fileLink.setAttribute("download", studentDocument.fileName);
+    document.body.appendChild(fileLink);
+    fileLink.click();
+    // After download, remove the element
+    fileLink.remove();
+  };
+
   return {
     getComponent,
     redrawComponent,
@@ -155,5 +174,6 @@ export function useFormioUtils() {
     setComponentValue,
     setRadioOptions,
     resetCheckBox,
+    downloadDocument,
   };
 }
